Change UI file names from Person to Applicant
