+++ src/main/resources/view/ApplicantListCard.fxml
